requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
awsebcli==3.20.10
beautifulsoup4==4.12.3
botocore==1.31.85
Brotli==1.1.0
cement==2.8.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.3
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
Django
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-bitfield==2.2.0
django-cors-headers==4.3.1
django-geo==0.8
django-heroku==0.3.1
django-js-asset==2.2.0
django-libs==2.0.3
django-mptt==0.16.0
django-ninja==1.1.0
django-payslip==0.3.2
django-separatedvaluesfield==0.6.0
django-tailwind
django-timezone-field==6.1.0
django-uuidfield==0.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
fonttools==4.48.1
gunicorn==21.2.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
lxml==5.1.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pathspec==0.10.1
pillow==10.2.0
platformdirs==4.2.0
postgis==1.0.4
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydyf==0.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyphen==0.14.0
pypiwin32
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.16.0
soupsieve==2.5
South==1.0.2
sqlparse==0.4.4
termcolor==1.1.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
unicodecsv==0.14.1
urllib3==1.26.18
views==0.3
virtualenv==20.25.0
wcwidth==0.1.9
weasyprint==61.0
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
